automation_platform/external_projects/departman_a/ocr_to_excel/requirements.txt.txt
pytesseract==0.3.10
Pillow==10.0.0
PyPDF2==3.0.1
pdf2image==1.16.3
pandas==2.0.3
openpyxl==3.1.2